requirements.txt
openai
streamlit
python-dotenv
faiss-cpu
tqdm
sentence-transformers
datasets
numpy
